+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased (????-??-??)
 
 * fix(stream): consider `ENOTFOUND` (macOS DNS error) as retryable for ApplicationLogStream
+* feat(stream): make `retryCount` public on ApplicationLogStream
 
 ## 8.0.3 (2024-01-30)