+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * fix(stream): consider `ENOTFOUND` (macOS DNS error) as retryable for ApplicationLogStream
 * feat(stream): make `retryCount` public on ApplicationLogStream
+* fix(stream): prevent uncaught errors on retry with ApplicationLogStream (firefox)
 
 ## 8.0.3 (2024-01-30)